requirements/base.txt
asgiref==3.5.2
attrs==22.1.0
awswrangler==2.17.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-guardian==2.4.0
django-polymorphic==3.1.0
djangorestframework==3.13.1
drf-yasg==1.21.3
Jinja2==3.1.2
loguru==0.6.0
numpy==1.23.3
pandas==1.5.0
pre-commit==2.15.0
psycopg2-binary==2.9.3
py==1.11.0
pyarrow==8.0.0
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-aws4auth==1.1.2
s3transfer==0.6.0
sqlparse==0.4.2
urllib3==1.26.12
matplotlib==3.6.0
matplotlib-inline==0.1.6
